refactor(desktop): extract workspace store from main process
